rename test_resilience_stats_LF.py to test_resilience_stats.py
